fix: 친구 삭제시 둘은 친구관계가 맞는지 확인 후 삭제 실행 하도록 변경했습니다.

+++ src/main/java/com/fiveman/newsfeed/comment/dto/CommentResponseDto.java
@@ -4,7 +4,6 @@ import com.fiveman.newsfeed.common.entity.Comment;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -15,7 +14,7 @@ public class CommentResponseDto {
     private Long userid;
     private Long commentId;
     private String content;
-    private Long LikeCount;
+    private Integer LikeCount;
     private LocalDateTime createAt;
     private LocalDateTime updatedAt;
